Let unit tests be independent of integration needs
We have a unit/ directory, but forgot to cleanly separate the integrations
tests into their own.  This manifested in some of the of fixtures required for
the integration tests to be utilized by the unit tests.  No more.
